requirements.txt
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
platformdirs==4.5.0
pyaml==25.7.0
Pygments==2.19.2
PyYAML==6.0.3
rich==14.2.0
textual==6.4.0
typing_extensions==4.15.0
uc-micro-py==1.0.3
